Extract ToolStripControlHost classes
