+++ src/System.Windows.Forms/src/System/Windows/Forms/ToolStripControlHost.cs
@@ -2,11 +2,8 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Collections;
 using System.ComponentModel;
-using System.ComponentModel.Design;
 using System.Drawing;
-using System.Runtime.InteropServices;
 using System.Windows.Forms.Layout;
 using static Interop;
 
@@ -15,7 +12,7 @@ namespace System.Windows.Forms
     /// <summary>
     ///  ToolStripItem that can host Controls.
     /// </summary>
-    public class ToolStripControlHost : ToolStripItem
+    public partial class ToolStripControlHost : ToolStripItem
     {
         private Control _control;
         private int _suspendSyncSizeCount;
@@ -944,280 +941,5 @@ namespace System.Windows.Forms
         /// </returns>
         protected override AccessibleObject CreateAccessibilityInstance()
             => new ToolStripControlHostAccessibleObject(this);
-
-        /// <summary>
-        ///  Defines the ToolStripControlHost AccessibleObject.
-        /// </summary>
-        [ComVisible(true)]
-        internal class ToolStripControlHostAccessibleObject : ToolStripItemAccessibleObject
-        {
-            private readonly ToolStripControlHost _ownerItem = null;
-
-            /// <summary>
-            ///  Initializes the new instance of ToolStripControlHostAccessibleObject.
-            /// </summary>
-            /// <param name="ownerItem">The owning ToolStripControlHost.</param>
-            public ToolStripControlHostAccessibleObject(ToolStripControlHost ownerItem) : base(ownerItem)
-            {
-                _ownerItem = ownerItem;
-            }
-
-            /// <summary>
-            ///  Gets a description of the default action for an object.
-            /// </summary>
-            public override string DefaultAction
-            {
-                get
-                {
-                    // Note: empty value is provided due to this Accessible object
-                    // represents the control container but not the contained control
-                    // itself.
-                    return string.Empty;
-                }
-            }
-
-            /// <summary>
-            ///  Performs the default action associated with this accessible object.
-            /// </summary>
-            public override void DoDefaultAction()
-            {
-                // Do nothing.
-            }
-
-            /// <summary>
-            ///  Gets the role of this accessible object.
-            /// </summary>
-            public override AccessibleRole Role
-            {
-                get
-                {
-                    AccessibleObject controlAccessibleObject  = _ownerItem.ControlAccessibilityObject;
-                    if (controlAccessibleObject != null)
-                    {
-                        return controlAccessibleObject.Role;
-                    }
-
-                    return AccessibleRole.Default;
-                }
-            }
-
-            /// <summary>
-            ///  Request to return the element in the specified direction.
-            /// </summary>
-            /// <param name="direction">Indicates the direction in which to navigate.</param>
-            /// <returns>Returns the element in the specified direction.</returns>
-            internal override UiaCore.IRawElementProviderFragment FragmentNavigate(UiaCore.NavigateDirection direction)
-            {
-                if (direction == UiaCore.NavigateDirection.FirstChild ||
-                    direction == UiaCore.NavigateDirection.LastChild)
-                {
-                    return _ownerItem.Control.AccessibilityObject;
-                }
-
-                // Handle Parent and other directions in base ToolStripItem.FragmentNavigate() method.
-                return base.FragmentNavigate(direction);
-            }
-
-            /// <summary>
-            ///  Return the element that is the root node of this fragment of UI.
-            /// </summary>
-            internal override UiaCore.IRawElementProviderFragmentRoot FragmentRoot
-                => _ownerItem.RootToolStrip.AccessibilityObject;
-        }
-
-        // Our implementation of ISite:
-        // Since the Control which is wrapped by ToolStripControlHost is a runtime instance, there is no way of knowing
-        // whether the control is in runtime or designtime.
-        // This implementation of ISite would be set to Control.Site when ToolStripControlHost.Site is set at DesignTime. (Refer to Site property on ToolStripControlHost)
-        // This implementation just returns the DesigMode property to be ToolStripControlHost's DesignMode property.
-        // Everything else is pretty much default implementation.
-        private class StubSite : ISite, IDictionaryService
-        {
-            private Hashtable _dictionary;
-            readonly IComponent _comp;
-            readonly IComponent _owner;
-
-            public StubSite(IComponent control, IComponent host)
-            {
-                _comp = control;
-                _owner = host;
-            }
-
-            /// <summary>
-            ///  When implemented by a class, gets the component associated with the <see cref='ISite'/>.
-            /// </summary>
-            IComponent ISite.Component => _comp;
-
-            /// <summary>
-            ///  When implemented by a class, gets the container associated with the <see cref='ISite'/>.
-            /// </summary>
-            IContainer ISite.Container => _owner.Site.Container;
-
-            /// <summary>
-            ///  When implemented by a class, determines whether the component is in design mode.
-            /// </summary>
-            bool ISite.DesignMode => _owner.Site.DesignMode;
-
-            /// <summary>
-            ///  When implemented by a class, gets or sets the name of
-            ///  the component associated with the <see cref='ISite'/>.
-            /// </summary>
-            string ISite.Name
-            {
-                get => _owner.Site.Name;
-                set => _owner.Site.Name = value;
-            }
-
-            /// <summary>
-            ///  Returns the requested service.
-            /// </summary>
-            object IServiceProvider.GetService(Type service)
-            {
-                if (service == null)
-                {
-                    throw new ArgumentNullException(nameof(service));
-                }
-
-                // We have to implement our own dictionary service. If we don't,
-                // the properties of the underlying component will end up being
-                // overwritten by our own properties when GetProperties is called
-                if (service == typeof(IDictionaryService))
-                {
-                    return this;
-                }
-
-                return _owner.Site?.GetService(service);
-            }
-
-            /// <summary>
-            ///  Retrieves the key corresponding to the given value.
-            /// </summary>
-            object IDictionaryService.GetKey(object value)
-            {
-                if (_dictionary != null)
-                {
-                    foreach (DictionaryEntry de in _dictionary)
-                    {
-                        object o = de.Value;
-                        if (value != null && value.Equals(o))
-                        {
-                            return de.Key;
-                        }
-                    }
-                }
-
-                return null;
-            }
-
-            /// <summary>
-            ///  Retrieves the value corresponding to the given key.
-            /// </summary>
-            object IDictionaryService.GetValue(object key)
-            {
-                if (_dictionary != null)
-                {
-                    return _dictionary[key];
-                }
-
-                return null;
-            }
-
-            /// <summary>
-            ///  Stores the given key-value pair in an object's site.  This key-value
-            ///  pair is stored on a per-object basis, and is a handy place to save
-            ///  additional information about a component.
-            /// </summary>
-            void IDictionaryService.SetValue(object key, object value)
-            {
-                if (_dictionary == null)
-                {
-                    _dictionary = new Hashtable();
-                }
-
-                if (value == null)
-                {
-                    _dictionary.Remove(key);
-                }
-                else
-                {
-                    _dictionary[key] = value;
-                }
-            }
-        }
-
-        /// <summary>
-        ///  Represents the ToolStrip hosted control accessible object which is responsible
-        ///  for accessible navigation within the ToolStrip standard items and hosted controls
-        ///  like TextBox, ComboBox, ProgressBar, etc.
-        /// </summary>
-        public class ToolStripHostedControlAccessibleObject : Control.ControlAccessibleObject
-        {
-            private ToolStripControlHost _toolStripControlHost;
-            private Control _toolStripHostedControl;
-
-            /// <summary>
-            ///  Creates the new instance of ToolStripHostedControlAccessibleObject.
-            /// </summary>
-            /// <param name="toolStripHostedControl">The ToolStrip control hosted in the ToolStripControlHost container.</param>
-            /// <param name="toolStripControlHost">The ToolStripControlHost container which hosts the control.</param>
-            public ToolStripHostedControlAccessibleObject(Control toolStripHostedControl, ToolStripControlHost toolStripControlHost) : base(toolStripHostedControl)
-            {
-                _toolStripControlHost = toolStripControlHost;
-                _toolStripHostedControl = toolStripHostedControl;
-            }
-
-            internal override UiaCore.IRawElementProviderFragmentRoot FragmentRoot
-            {
-                get
-                {
-                    if (_toolStripHostedControl != null // Hosted control should not be null.
-                        && _toolStripControlHost != null // ToolStripControlHost is a container for ToolStripControl.
-                        && _toolStripControlHost.Owner != null) // Owner is the ToolStrip.
-                    {
-                        return _toolStripControlHost.Owner.AccessibilityObject; 
-                    }
-
-                    return base.FragmentRoot;
-                }
-            }
-
-            internal override UiaCore.IRawElementProviderFragment FragmentNavigate(UiaCore.NavigateDirection direction)
-            {
-                if (_toolStripHostedControl != null &&
-                    _toolStripControlHost != null)
-                {
-                    switch (direction)
-                    {
-                        case UiaCore.NavigateDirection.Parent:
-                        case UiaCore.NavigateDirection.PreviousSibling:
-                        case UiaCore.NavigateDirection.NextSibling:
-                            return _toolStripControlHost.AccessibilityObject.FragmentNavigate(direction);
-                    }
-                }
-
-                return base.FragmentNavigate(direction);
-            }
-
-            internal override object GetPropertyValue(UiaCore.UIA propertyID)
-            {
-                switch (propertyID)
-                {
-                    case UiaCore.UIA.HasKeyboardFocusPropertyId:
-                        return (State & AccessibleStates.Focused) == AccessibleStates.Focused;
-                }
-
-                return base.GetPropertyValue(propertyID);
-            }
-
-            internal override bool IsPatternSupported(UiaCore.UIA patternId)
-            {
-                if (patternId == UiaCore.UIA.ValuePatternId)
-                {
-                    return true;
-                }
-
-                return base.IsPatternSupported(patternId);
-            }
-        }
     }
 }
